requirements.txt
aiofiles==24.1.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
fastapi==0.111.1
mypy==1.9.0
passlib==1.7.4
playwright==1.45.1
psycopg2-binary==2.9.9
pydantic_settings==2.3.4
pydocstyle==6.3.0
pyjwt==2.8.0
pylint==3.2.5
pytest-asyncio==0.23.8
pytest-cov==5.0.0
pytest-env==1.1.3
pytest==8.3.2
python-dotenv==1.0.1
redis==5.0.8
sqlalchemy==2.0.31
types-aiofiles==24.1.0.20240626
types-beautifulsoup4==4.12.0.20240511
types-html5lib==1.1.11.20240228
celery==5.4.0